third_party/blink/web_tests/external/wpt/css/css-text/inheritance-expected.txt

This is a testharness.js-based test.
[FAIL] Property hanging-punctuation has initial value none
  assert_true: hanging-punctuation doesn't seem to be supported in the computed style expected true got false
[FAIL] Property hanging-punctuation inherits
  assert_true: hanging-punctuation doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-align-all has initial value start
  assert_true: text-align-all doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-align-all inherits
  assert_true: text-align-all doesn't seem to be supported in the computed style expected true got false
Harness: the test ran to completion.
